refacotr(serializers): split serializers in different files
