+++ core/src/actors/session/handlers.rs
@@ -29,7 +29,7 @@ use super::{
 use witnet_data_structures::{
     builders::from_address,
     chain::{Block, CheckpointBeacon, Hash, InventoryEntry, InventoryItem},
-    serializers::TryFrom,
+    serializers::decoders::TryFrom,
     types::{
         Address, Command, InventoryAnnouncement, InventoryRequest, LastBeacon,
         Message as WitnetMessage, Peers, Version,
