+++ data_structures/src/serializers/mod.rs
@@ -1,0 +1,5 @@
+/// Tools to convert from flatbuffers into rust structures
+pub mod decoders;
+
+/// Tools to convert from rust structures into flatbuffers
+pub mod encoders;
